caching first version
The caching method has been added. However, since the code stores all the predictions (a torch array with a length equal to the number of nodes), the resulting files are extremely large (372GB). Therefore, it would be better to only store the top k results per query.

Additionally, I had to split the list of all possible 1p queries into smaller chunks because processing all of them at once causes an out-of-memory error.